HBASE-22676 Move all the code in hbase-rsgroup to hbase-server and remove hbase-rsgroup module
